--- plugins/woocommerce-blocks/assets/js/blocks/add-to-cart-with-options/quantity-selector/components/quantity-stepper.tsx
@@ -1,34 +1,0 @@
-export const QuantityStepper = ( {
-	isSiteEditor,
-}: {
-	isSiteEditor: boolean;
-} ) => {
-	return (
-		<div className="quantity wc-block-components-quantity-selector">
-			<button className="wc-block-components-quantity-selector__button wc-block-components-quantity-selector__button--minus">
-				-
-			</button>
-			<input
-				style={
-					// In the post editor, the editor isn't in an iframe, so WordPress styles are applied. We need to remove them.
-					! isSiteEditor
-						? {
-								backgroundColor: '#ffffff',
-								lineHeight: 'normal',
-								minHeight: 'unset',
-								boxSizing: 'unset',
-								borderRadius: 'unset',
-						  }
-						: {}
-				}
-				type="number"
-				value="1"
-				className="input-text qty text"
-				readOnly
-			/>
-			<button className="wc-block-components-quantity-selector__button wc-block-components-quantity-selector__button--plus">
-				+
-			</button>
-		</div>
-	);
-};
