+++ plugins/woocommerce/src/Blocks/BlockTypes/AddToCartWithOptionsQuantitySelector.php
@@ -17,6 +17,22 @@ class AddToCartWithOptionsQuantitySelector extends AbstractBlock {
 	protected $block_name = 'add-to-cart-with-options-quantity-selector';
 
 	/**
+	 * Get the block's attributes.
+	 *
+	 * @param array $attributes Block attributes. Default empty array.
+	 * @return array  Block attributes merged with defaults.
+	 */
+	private function parse_attributes( $attributes ) {
+		// These should match what's set in JS `registerBlockType`.
+		$defaults = array(
+			'quantitySelectorStyle' => 'input',
+		);
+
+		return wp_parse_args( $attributes, $defaults );
+	}
+
+
+	/**
 	 * Enqueue assets specific to this block.
 	 * We enqueue frontend scripts only if the quantitySelectorStyle is set to 'stepper'.
 	 *
@@ -39,7 +55,7 @@ class AddToCartWithOptionsQuantitySelector extends AbstractBlock {
 	 * @param string $product_name Product name.
 	 * @return stringa Quantity input HTML with increment and decrement buttons.
 	 */
-	private static function add_steppers( $product_html, $product_name ) {
+	private function add_steppers( $product_html, $product_name ) {
 		// Regex pattern to match the <input> element with id starting with 'quantity_'.
 		$pattern = '/(<input[^>]*id="quantity_[^"]*"[^>]*\/>)/';
 		// Replacement string to add button BEFORE the matched <input> element.
@@ -60,7 +76,7 @@ class AddToCartWithOptionsQuantitySelector extends AbstractBlock {
 	 *
 	 * @return string The Quantity Selector HTML with classes added.
 	 */
-	private static function add_stepper_classes( $product_html ) {
+	private function add_stepper_classes( $product_html ) {
 		$html = new \WP_HTML_Tag_Processor( $product_html );
 
 		// Add classes to the form.
@@ -76,14 +92,30 @@ class AddToCartWithOptionsQuantitySelector extends AbstractBlock {
 		return $html->get_updated_html();
 	}
 
+
 	/**
-	 * Get the quantity selector markup.
+	 * Render the block.
 	 *
-	 * @param WC_Product $product The product object.
-	 * @param array      $attributes The attributes of the block.
-	 * @return string The quantity selector markup.
+	 * @param array    $attributes Block attributes.
+	 * @param string   $content Block content.
+	 * @param WP_Block $block Block instance.
+	 *
+	 * @return string | void Rendered block output.
 	 */
-	public static function get_quantity_selector_markup( $product, $attributes ) {
+	protected function render( $attributes, $content, $block ) {
+		global $product;
+		$previous_product = $product;
+
+		// Try to load the product from the block context, if not available,
+		// use the global $product.
+		$post_id = isset( $block->context['postId'] ) ? $block->context['postId'] : '';
+		$post    = $post_id ? wc_get_product( $post_id ) : null;
+		if ( $post instanceof \WC_Product ) {
+			$product = $post;
+		} elseif ( ! $product instanceof \WC_Product ) {
+			return '';
+		}
+
 		$is_external_product_with_url        = $product instanceof \WC_Product_External && $product->get_product_url();
 		$can_only_be_purchased_one_at_a_time = $product->is_sold_individually();
 
@@ -120,9 +152,10 @@ class AddToCartWithOptionsQuantitySelector extends AbstractBlock {
 		$product_html = ob_get_clean();
 
 		$product_name = $product->get_name();
-		$product_html = $is_stepper_style ? self::add_steppers( $product_html, $product_name ) : $product_html;
+		$product_html = $is_stepper_style ? $this->add_steppers( $product_html, $product_name ) : $product_html;
 
-		$product_html       = $is_stepper_style ? self::add_stepper_classes( $product_html ) : $product_html;
+		$parsed_attributes  = $this->parse_attributes( $attributes );
+		$product_html       = $is_stepper_style ? $this->add_stepper_classes( $product_html ) : $product_html;
 		$classes_and_styles = StyleAttributesUtils::get_classes_and_styles_by_attributes( $attributes, array(), array( 'extra_classes' ) );
 
 		$classes = implode(
@@ -143,7 +176,7 @@ class AddToCartWithOptionsQuantitySelector extends AbstractBlock {
 			)
 		);
 
-		$markup = sprintf(
+		$form = sprintf(
 			'<div %1$s %2$s>%3$s</div>',
 			$wrapper_attributes,
 			$is_stepper_style ? 'data-wc-interactive=\'' . wp_json_encode(
@@ -155,36 +188,8 @@ class AddToCartWithOptionsQuantitySelector extends AbstractBlock {
 			$product_html
 		);
 
-		return $markup;
-	}
-
-	/**
-	 * Render the block.
-	 *
-	 * @param array    $attributes Block attributes.
-	 * @param string   $content Block content.
-	 * @param WP_Block $block Block instance.
-	 *
-	 * @return string | void Rendered block output.
-	 */
-	protected function render( $attributes, $content, $block ) {
-		global $product;
-		$previous_product = $product;
-
-		// Try to load the product from the block context, if not available,
-		// use the global $product.
-		$post_id = isset( $block->context['postId'] ) ? $block->context['postId'] : '';
-		$post    = $post_id ? wc_get_product( $post_id ) : null;
-		if ( $post instanceof \WC_Product ) {
-			$product = $post;
-		} elseif ( ! $product instanceof \WC_Product ) {
-			return '';
-		}
-
-		$markup = $this->get_quantity_selector_markup( $product, $attributes );
-
 		$product = $previous_product;
 
-		return $markup;
+		return $form;
 	}
 }
