+++ plugins/woocommerce/tests/php/src/Blocks/BlockTypes/AddToCartWithOptions.php
@@ -15,19 +15,31 @@ use Automattic\WooCommerce\Tests\Blocks\Mocks\AddToCartWithOptionsGroupedProduct
  * Tests for the AddToCartWithOptions block type
  */
 class AddToCartWithOptions extends \WP_UnitTestCase {
+
+	/**
+	 * Tracks whether blocks have been registered.
+	 *
+	 * @var bool
+	 */
+	protected static $blocks_registered = false;
+
 	/**
 	 * Initiate the mock object.
 	 */
 	protected function setUp(): void {
 		parent::setUp();
 
-		// We need to register the blocks after set up. They are no registered
-		// on `init` because `init` is called with a classic theme.
-		new AddToCartWithOptionsMock();
-		new AddToCartWithOptionsQuantitySelectorMock();
-		new AddToCartWithOptionsGroupedProductSelectorMock();
-		new AddToCartWithOptionsGroupedProductSelectorItemCTAMock();
-		new AddToCartWithOptionsGroupedProductSelectorItemTemplateMock();
+		if ( ! self::$blocks_registered ) {
+			// We need to register the blocks after set up. They are no registered
+			// on `init` because `init` is called with a classic theme.
+			new AddToCartWithOptionsMock();
+			new AddToCartWithOptionsQuantitySelectorMock();
+			new AddToCartWithOptionsGroupedProductSelectorMock();
+			new AddToCartWithOptionsGroupedProductSelectorItemTemplateMock();
+			new AddToCartWithOptionsGroupedProductSelectorItemCTAMock();
+
+			self::$blocks_registered = true;
+		}
 	}
 
 	/**
@@ -42,67 +54,62 @@ class AddToCartWithOptions extends \WP_UnitTestCase {
 	/**
 	 * Tests that the correct content is rendered for each product type.
 	 */
-	// public function test_product_type_add_to_cart_render() {
-	// add_action( 'woocommerce_custom_add_to_cart', array( $this, 'print_custom_product_type_add_to_cart_markup' ) );
+	public function test_product_type_add_to_cart_render() {
+		add_action( 'woocommerce_custom_add_to_cart', array( $this, 'print_custom_product_type_add_to_cart_markup' ) );
 
-	// global $product;
-	// $product = new \WC_Product_Simple();
-	// $product->set_regular_price( 10 );
-	// $product_id = $product->save();
-	// $markup     = do_blocks( '<!-- wp:woocommerce/single-product {"productId":' . $product_id . '} --><!-- wp:woocommerce/add-to-cart-with-options /--><!-- /wp:woocommerce/single-product -->' );
+		global $product;
+		$product = new \WC_Product_Simple();
+		$product->set_regular_price( 10 );
+		$product_id = $product->save();
+		$markup     = do_blocks( '<!-- wp:woocommerce/single-product {"productId":' . $product_id . '} --><!-- wp:woocommerce/add-to-cart-with-options /--><!-- /wp:woocommerce/single-product -->' );
 
-	// Single Products contain the Add to Cart button and the quantity selector blocks.
-	// $this->assertStringContainsString( 'wp-block-woocommerce-product-button', $markup, 'The Simple Product Add to Cart with Options contains the product button block.' );
-	// $this->assertStringContainsString( 'woocommerce/add-to-cart-with-options-quantity-selector', $markup, 'The Simple Product Add to Cart with Options contains the quantity selector block.' );
+		// Single Products contain the Add to Cart button and the quantity selector blocks.
+		$this->assertStringContainsString( 'wp-block-woocommerce-product-button', $markup, 'The Simple Product Add to Cart with Options contains the product button block.' );
+		$this->assertStringContainsString( 'woocommerce/add-to-cart-with-options-quantity-selector', $markup, 'The Simple Product Add to Cart with Options contains the quantity selector block.' );
 
-	// $product    = new \WC_Product_External();
-	// $product_id = $product->save();
-	// $markup     = do_blocks( '<!-- wp:woocommerce/single-product {"productId":' . $product_id . '} --><!-- wp:woocommerce/add-to-cart-with-options /--><!-- /wp:woocommerce/single-product -->' );
+		$product    = new \WC_Product_External();
+		$product_id = $product->save();
+		$markup     = do_blocks( '<!-- wp:woocommerce/single-product {"productId":' . $product_id . '} --><!-- wp:woocommerce/add-to-cart-with-options /--><!-- /wp:woocommerce/single-product -->' );
 
-	// External Products contain the Add to Cart button block but do not contain the quantity selector block.
-	// $this->assertStringContainsString( 'wp-block-woocommerce-product-button', $markup, 'The External Product Add to Cart with Options contains the product button block.' );
-	// $this->assertStringNotContainsString( 'woocommerce/add-to-cart-with-options-quantity-selector', $markup, 'The External Product Add to Cart with Options does not contain the quantity selector block.' );
+		// External Products contain the Add to Cart button block but do not contain the quantity selector block.
+		$this->assertStringContainsString( 'wp-block-woocommerce-product-button', $markup, 'The External Product Add to Cart with Options contains the product button block.' );
+		$this->assertStringNotContainsString( 'woocommerce/add-to-cart-with-options-quantity-selector', $markup, 'The External Product Add to Cart with Options does not contain the quantity selector block.' );
 
-	// $product    = new WC_Product_Custom();
-	// $product_id = $product->save();
-	// $markup     = do_blocks( '<!-- wp:woocommerce/single-product {"productId":' . $product_id . '} --><!-- wp:woocommerce/add-to-cart-with-options /--><!-- /wp:woocommerce/single-product -->' );
+		$product    = new WC_Product_Custom();
+		$product_id = $product->save();
+		$markup     = do_blocks( '<!-- wp:woocommerce/single-product {"productId":' . $product_id . '} --><!-- wp:woocommerce/add-to-cart-with-options /--><!-- /wp:woocommerce/single-product -->' );
 
-	// Third-party product types use their own template.
-	// $this->assertStringContainsString( 'Custom Product Type Add to Cart Form', $markup, 'The Custom Product Type Add to Cart with Options contains the custom product type add to cart form.' );
+		// Third-party product types use their own template.
+		$this->assertStringContainsString( 'Custom Product Type Add to Cart Form', $markup, 'The Custom Product Type Add to Cart with Options contains the custom product type add to cart form.' );
 
-	// remove_action( 'woocommerce_custom_add_to_cart', array( $this, 'print_custom_product_type_add_to_cart_markup' ) );
-	// }
+		remove_action( 'woocommerce_custom_add_to_cart', array( $this, 'print_custom_product_type_add_to_cart_markup' ) );
+	}
 
 	/**
 	 * Tests that the correct CTA is rendered in the Grouped Product Selector.
 	 */
 	public function test_grouped_product_selector_cta() {
-		// Create a grouped product.
-		// Create a simple product.
-		// Create a simple product sold individually.
-		// Create a variable product.
-		// $simple_product->set_sold_individually( true );
-
-		// First test: make the simple product to be a child of the grouped product.
-		// Test that we display a quantity selector.
 		$simple_product = new \WC_Product_Simple();
-		$simple_product->set_price( 10 );
+		$simple_product->set_regular_price( 10 );
 		$simple_product_id = $simple_product->save();
 		$grouped_product   = new \WC_Product_Grouped();
-		// $grouped_product->set_children( array( $simple_product_id ) );
+		$grouped_product->set_children( array( $simple_product_id ) );
 		$grouped_product_id = $grouped_product->save();
+
 		global $product;
 		$product = $grouped_product;
-		echo 'GROUPED PRODUCT ID: ' . $grouped_product_id . '<br>';
+
 		$markup = do_blocks( '<!-- wp:woocommerce/single-product {"productId":' . $grouped_product_id . '} --><!-- wp:woocommerce/add-to-cart-with-options /--><!-- /wp:woocommerce/single-product -->' );
-		$this->assertStringContainsString( 'wc-block-add-to-cart-with-options-grouped-product-selector-item-cta', $markup, 'The Grouped Product Add to Cart with Options form contains a checkbox.' );
-		// $this->assertStringContainsString( '<input type="number', $markup, 'The Grouped Product Add to Cart with Options form contains a checkbox.' );
+		$this->assertStringContainsString( 'type="number"', $markup, 'The Grouped Product Add to Cart with Options form contains a checkbox.' );
 
-		// Second test: make the simple product sold individually.
-		// Test that we display a checkbox.
-		// $this->assertStringContainsString( '<input type="checkbox', $markup, 'The Grouped Product Add to Cart with Options form contains a checkbox.' );
+		$simple_product->set_sold_individually( true );
+		$simple_product->save();
+		$markup = do_blocks( '<!-- wp:woocommerce/single-product {"productId":' . $grouped_product_id . '} --><!-- wp:woocommerce/add-to-cart-with-options /--><!-- /wp:woocommerce/single-product -->' );
+		$this->assertStringContainsString( 'type="checkbox"', $markup, 'The Grouped Product Add to Cart with Options form contains a checkbox.' );
 
-		// Third test: make the variable product to be a child of the grouped product.
-		// Test that we display a button.
+		$simple_product->set_stock_status( 'outofstock' );
+		$simple_product->save();
+		$markup = do_blocks( '<!-- wp:woocommerce/single-product {"productId":' . $grouped_product_id . '} --><!-- wp:woocommerce/add-to-cart-with-options /--><!-- /wp:woocommerce/single-product -->' );
+		$this->assertStringContainsString( 'Read more', $markup, 'The Grouped Product Add to Cart with Options form contains a button.' );
 	}
 }
